requirements.txt
pytest~=7.3.1
eyes-selenium==5.12.0